pyproject.toml
[tool.poetry]
name = "s-bert-for-ja"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sentence-transformers = "^2.2.2"
logzero = "^1.7.0"
tqdm = "^4.66.1"
pydantic = "^2.5.3"
fugashi = "^1.3.0"
unidic-lite = "^1.0.8"
openai = "^1.6.1"
python-dotenv = "^1.0.0"
tiktoken = "^0.5.2"
polars = "^0.20.3"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.9"
mypy = "^1.8.0"
pytest = "^7.4.4"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
show_error_context = true
show_column_numbers = true
ignore_missing_imports = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
check_untyped_defs = true

[tool.ruff]
exclude = [
    ".venv",
    ".vscode",
    "data",
    "notebook",
    "__pycache__",
    ".git",
]

line-length = 120
indent-width = 4

[tool.ruff.lint]
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"